Moves elm package and source directory management out of builder into dependency manager
