+++ lib/builder.ts
@@ -1,27 +1,9 @@
 import * as Bluebird from "bluebird";
-import * as _ from "lodash";
 import chalk from "chalk";
 import * as path from "path";
-import * as shelljs from "shelljs";
 import * as childProcess from "child_process";
-import * as promptly from "promptly";
-
 import {createLogger, Logger} from "./logger";
-import {Dependencies, ExecutionContext, LoboConfig, PluginTestFrameworkWithConfig, Reject, Resolve} from "./plugin";
-import {createElmPackageHelper, ElmPackageHelper} from "./elm-package-helper";
-import {createUtil, Util} from "./util";
-
-
-interface ElmPackageJson {
-  dependencies: Dependencies;
-  sourceDirectories: string[];
-}
-
-interface ElmPackageCompare {
-  readonly base: ElmPackageJson;
-  readonly test: ElmPackageJson;
-}
-
+import {ExecutionContext, LoboConfig, Reject, Resolve} from "./plugin";
 
 export interface Builder {
   build(context: ExecutionContext): Bluebird<ExecutionContext>;
@@ -29,298 +11,15 @@ export interface Builder {
 
 export class BuilderImp implements Builder {
 
-  private logger: Logger;
-  private elmPackageHelper: ElmPackageHelper;
-  private util: Util;
-  private yOrN: string = chalk.dim(" [Y/n]");
+  private readonly logger: Logger;
 
-  constructor(elmPackageHelper: ElmPackageHelper, logger: Logger, util: Util) {
-    this.elmPackageHelper = elmPackageHelper;
+  constructor(logger: Logger) {
     this.logger = logger;
-    this.util = util;
   }
 
   public build(context: ExecutionContext): Bluebird<ExecutionContext> {
-    this.logger.info("-----------------------------------[ BUILD ]------------------------------------");
-
-    let baseElmPackageDir = ".";
-    let testElmPackageDir = context.testDirectory;
-    let testDir = path.dirname(context.testFile);
-    let steps: Array<() => Bluebird<void>> = [];
-
-    if (context.config.noUpdate) {
-      this.logger.info("Ignored sync of base and test elm-package.json files due to configuration");
-    } else {
-      steps = steps.concat([
-        () => this.ensureElmPackageExists(context.config, baseElmPackageDir, "current"),
-        () => this.ensureElmPackageExists(context.config, testElmPackageDir, "tests"),
-        () => this.syncTestElmPackage(context.config, baseElmPackageDir, testElmPackageDir, testDir)]);
-    }
-
-    steps = steps.concat([
-      () => this.installDependencies(context.config, context.testDirectory),
-      () => this.make(context.config, context.buildOutputFilePath, context.testDirectory, context.testFile)
-    ]);
-
-    return Bluebird.mapSeries(steps, (item: () => Bluebird<ExecutionContext>) => item())
-      .return(context);
-  }
-
-  public ensureElmPackageExists(config: LoboConfig, elmPackageDir: string, location: string): Bluebird<void> {
-    return new Bluebird((resolve: Resolve<void>, reject: Reject) => {
-      if (shelljs.test("-e", this.elmPackageHelper.path(elmPackageDir))) {
-        resolve();
-        return;
-      }
-
-      if (!config.prompt) {
-        this.runElmPackageInstall(config, elmPackageDir, false, resolve, reject);
-        return;
-      }
-
-      promptly.confirm(
-        "Unable to find elm-package.json in the " + location + " directory" +
-        "\n\nMay I create a minimal elm-package.json for you?" + this.yOrN,
-        {"default": "yes"},
-        (err, value) => {
-          if (err) {
-            reject(err);
-          } else if (value && value.toString() === "true") {
-            this.runElmPackageInstall(config, elmPackageDir, true, resolve, reject);
-          } else {
-            reject();
-          }
-        });
-    });
-  }
-
-  public syncTestElmPackage(config: LoboConfig, baseElmPackageDir: string, testElmPackageDir: string, testDir: string): Bluebird<void> {
-    let steps = <Array<(result: ElmPackageCompare) => Bluebird<ElmPackageCompare>>> [
-      () => this.readElmPackage(baseElmPackageDir, testElmPackageDir),
-      (result: ElmPackageCompare) =>
-        this.updateSourceDirectories(config, baseElmPackageDir, result.base, testElmPackageDir, testDir, result.test),
-      (result: ElmPackageCompare) => this.updateDependencies(config, result.base, testElmPackageDir, result.test)];
-
-    let value: ElmPackageCompare;
-
-    return Bluebird
-      .mapSeries(steps, (item: (result: ElmPackageCompare) => Bluebird<ElmPackageCompare>) => item(value)
-        .then((result: ElmPackageCompare) => value = result))
-      .return();
-  }
-
-  public readElmPackage(baseElmPackageDir: string, testElmPackageDir: string): Bluebird<ElmPackageCompare> {
-    return new Bluebird<ElmPackageCompare>((resolve: Resolve<ElmPackageCompare>, reject: Reject) => {
-      let baseElmPackage = this.elmPackageHelper.read(baseElmPackageDir);
-
-      if (!baseElmPackage) {
-        this.logger.error("Unable to read the main elm-package.json file. Please check that is a valid json file");
-        reject();
-      }
-
-      let testElmPackage = this.elmPackageHelper.read(testElmPackageDir);
-
-      if (!testElmPackage) {
-        this.logger.error("Unable to read the test elm-package.json file. Please check that is a valid json file");
-        reject();
-      }
-
-      resolve(<ElmPackageCompare> {base: baseElmPackage, test: testElmPackage});
-    });
-  }
-
-  public updateSourceDirectories(config: LoboConfig, baseElmPackageDir: string, baseElmPackage: ElmPackageJson, testElmPackageDir: string,
-                                 testDir: string, testElmPackage: ElmPackageJson): Bluebird<ElmPackageCompare> {
-    return new Bluebird<ElmPackageCompare>((resolve: Resolve<ElmPackageCompare>, reject: Reject) => {
-      let sourceDirectories =
-        this.mergeSourceDirectories(baseElmPackage, baseElmPackageDir, testElmPackage, testElmPackageDir, testDir, config.testFramework);
-      let diff = _.difference(sourceDirectories, testElmPackage.sourceDirectories);
-
-      if (diff.length === 0) {
-        resolve({base: baseElmPackage, test: testElmPackage});
-        return;
-      }
-
-      if (!config.prompt) {
-        testElmPackage = this.updateSourceDirectoriesAction(sourceDirectories, testElmPackageDir, testElmPackage);
-        resolve({base: baseElmPackage, test: testElmPackage});
-        return;
-      }
-
-      promptly.confirm(
-        "The source-directories of the test elm-package.json needs to be updated to " +
-        "contain:\n" + diff.join("\n") + "\n\nMay I add them to elm-package.json for you?" +
-        this.yOrN,
-        {"default": "yes"},
-        (err, value) => {
-          if (err) {
-            reject(err);
-          } else if (value && value.toString() === "true") {
-            testElmPackage = this.updateSourceDirectoriesAction(sourceDirectories, testElmPackageDir, testElmPackage);
-            resolve({base: baseElmPackage, test: testElmPackage});
-          } else {
-            reject();
-          }
-        });
-    });
-  }
-
-  public updateSourceDirectoriesAction(sourceDirectories: string[], testElmPackageDir: string, testElmPackage: ElmPackageJson):
-  ElmPackageJson {
-    testElmPackage.sourceDirectories = sourceDirectories;
-    this.elmPackageHelper.write(testElmPackageDir, testElmPackage);
-
-    return testElmPackage;
-  }
-
-  public updateDependencies(config: LoboConfig, baseElmPackage: ElmPackageJson, testElmPackageDir: string,
-                            testElmPackage: ElmPackageJson): Bluebird<ElmPackageCompare> {
-    return new Bluebird<ElmPackageCompare>((resolve: Resolve<ElmPackageCompare>, reject: Reject) => {
-      let dependencies = this.mergeDependencies(baseElmPackage, testElmPackage, config.testFramework);
-      let existing = _.toPairs(testElmPackage.dependencies);
-      let diff = _.filter(dependencies, base => this.isNotExistingDependency(existing, base));
-
-      if (diff.length === 0) {
-        resolve({base: baseElmPackage, test: testElmPackage});
-        return;
-      }
-
-      let diffString = _.map(diff, kp => kp[0] + ": " + kp[1]);
-
-      if (!config.prompt) {
-        testElmPackage = this.updateDependenciesAction(dependencies, testElmPackageDir, testElmPackage);
-        resolve({base: baseElmPackage, test: testElmPackage});
-        return;
-      }
-
-      promptly.confirm(
-        "The dependencies of the test elm-package.json need to be updated to contain:\n" +
-        diffString.join("\n") + "\n\nMay I add them to elm-package.json for you--------?" +
-        this.yOrN,
-        {"default": "yes"}, (err, value) => {
-          if (err) {
-            reject(err);
-          } else if (value && value.toString() === "true") {
-            testElmPackage = this.updateDependenciesAction(dependencies, testElmPackageDir, testElmPackage);
-            resolve({base: baseElmPackage, test: testElmPackage});
-          } else {
-            reject();
-          }
-        });
-    });
-  }
-
-  public updateDependenciesAction(dependencies: string[][], testElmPackageDir: string, testElmPackage: ElmPackageJson): ElmPackageJson {
-    let sortedDependencies: string[][] = _.sortBy(dependencies, (kp: string[]) => kp[0]);
-    testElmPackage.dependencies = _.fromPairs(sortedDependencies);
-    this.elmPackageHelper.write(testElmPackageDir, testElmPackage);
-
-    return testElmPackage;
-  }
-
-  public mergeSourceDirectories(baseElmPackage: ElmPackageJson, baseElmPackageDir: string, testElmPackage: ElmPackageJson,
-                                testElmPackageDir: string, testDir: string, testFramework: PluginTestFrameworkWithConfig): string[] {
-    let sourceDirs: string[] = _.clone(testElmPackage.sourceDirectories);
-
-    if (!sourceDirs) {
-      sourceDirs = [];
-    }
-
-    if (sourceDirs.indexOf(".") === -1) {
-      sourceDirs.push(".");
-    }
-
-    if (sourceDirs.indexOf(testDir) === -1) {
-      sourceDirs.push(testDir);
-    }
-
-    sourceDirs = this.addSourceDirectories(baseElmPackage.sourceDirectories, baseElmPackageDir, testElmPackageDir, sourceDirs);
-
-    let dirPath = path.join(__dirname, "..");
-    sourceDirs = this.addSourceDirectories(testFramework.config.sourceDirectories, dirPath, testElmPackageDir, sourceDirs);
-
-    return sourceDirs;
-  }
-
-  public addSourceDirectories(additions: string[], additionDir: string, testElmPackageDir: string, sourceDirs: string[]): string[] {
-    if (!additions) {
-      return sourceDirs;
-    }
-
-    let relativePath = path.relative(testElmPackageDir, additionDir);
-    let absoluteSourceDirs = _.map(sourceDirs, p => this.util.resolveDir(testElmPackageDir, p));
-
-    let relativeSourceDirectories =
-      _.map(additions, p => path.join(relativePath, p)
-        .replace(/\\/g, "/"))
-        .filter(p => absoluteSourceDirs.indexOf(this.util.resolveDir(testElmPackageDir, p)) === -1);
-
-    return sourceDirs.concat(relativeSourceDirectories);
-  }
-
-  public mergeDependencies(baseElmPackage: ElmPackageJson, testElmPackage: ElmPackageJson,
-                           testFramework: PluginTestFrameworkWithConfig): string[][] {
-    let dependencies: string[][] = _.toPairs(testElmPackage.dependencies);
-
-    if (baseElmPackage.dependencies) {
-      let baseDependencies: string[][] = _.toPairs(baseElmPackage.dependencies)
-        .filter((base: string[]) => this.isNotExistingDependency(dependencies, base));
-
-      dependencies = dependencies.concat(baseDependencies);
-    }
-
-    if (testFramework.config.dependencies) {
-      let testFrameworkDependencies = _.toPairs(testFramework.config.dependencies)
-        .filter((base) => this.isNotExistingDependency(dependencies, base));
-
-      dependencies = dependencies.concat(testFrameworkDependencies);
-    }
-
-    return dependencies;
-  }
-
-  public isNotExistingDependency(dependencies: string[][], candidate: string[]): boolean {
-    return !_.find(dependencies, x => {
-      return x[0] === candidate[0] && !this.elmPackageHelper.isImprovedMinimumConstraint(x[1], candidate[1]);
-    });
-  }
-
-  public installDependencies(config: LoboConfig, testDirectory: string): Bluebird<void> {
-    return new Bluebird((resolve: Resolve<void>, reject: Reject) => {
-      this.runElmPackageInstall(config, testDirectory, config.prompt, resolve, reject);
-    });
-  }
-
-  public runElmPackageInstall(config: LoboConfig, directory: string, prompt: boolean, resolve: Resolve<void>, reject: Reject): void {
-    if (config.noInstall) {
-      this.logger.info("Ignored running of elm-package install due to configuration");
-      resolve();
-      return;
-    }
-
-    let command = "elm-package";
-
-    if (config.compiler) {
-      command = path.join(config.compiler, command);
-    }
-
-    command += " install";
-
-    if (!prompt) {
-      command += " --yes";
-    }
-
-    try {
-      // run as child process using current process stdio so that colored output is returned
-      let options = {cwd: directory, stdio: [process.stdin, process.stdout, process.stderr]};
-      this.logger.trace(command);
-      childProcess.execSync(command, options);
-      resolve();
-    } catch (err) {
-      this.logger.debug("elm package install failed in the test directory");
-      this.logger.debug(err);
-      reject(err);
-    }
+    return this.make(context.config, context.buildOutputFilePath, context.testDirectory, context.testFile)
+      .then(() => context);
   }
 
   public make(config: LoboConfig, buildOutputFilePath: string, testDirectory: string, testFile: string): Bluebird<void> {
@@ -365,5 +64,5 @@ export class BuilderImp implements Builder {
 }
 
 export function createBuilder(): Builder {
-  return new BuilderImp(createElmPackageHelper(), createLogger(), createUtil());
+  return new BuilderImp(createLogger());
 }
