+++ lib/elm-package-helper.ts
@@ -4,6 +4,7 @@ import * as _ from "lodash";
 
 import {createLogger, Logger} from "./logger";
 import {createUtil, Util} from "./util";
+import {PluginTestFrameworkWithConfig} from "./plugin";
 
 export interface Dependencies {
   [index: string]: string;
@@ -16,6 +17,11 @@ export interface ElmPackageJson {
 
 export interface ElmPackageHelper {
   isImprovedMinimumConstraint(dependency: string, candidate: string): boolean;
+  isNotExistingDependency(dependencies: string[][], candidate: string[]): boolean;
+  mergeDependencies(baseElmPackage: ElmPackageJson, testElmPackage: ElmPackageJson, testFramework: PluginTestFrameworkWithConfig)
+    : string[][];
+  mergeSourceDirectories(baseElmPackage: ElmPackageJson, baseElmPackageDir: string, testElmPackage: ElmPackageJson,
+                         testElmPackageDir: string, testDir: string): string[];
   path(elmPackageJsonDirectory: string): string;
   read(elmPackageJsonDirectory: string): ElmPackageJson | undefined;
   write(elmPackageJsonDirectory: string, elmPackage: ElmPackageJson): void;
@@ -29,6 +35,22 @@ export class ElmPackageHelperImp implements ElmPackageHelper {
   constructor(logger: Logger, util: Util) {
     this.logger = logger;
     this.util = util;
+  }
+
+  public addSourceDirectories(additions: string[], additionDir: string, testElmPackageDir: string, sourceDirs: string[]): string[] {
+    if (!additions) {
+      return sourceDirs;
+    }
+
+    let relativePath = path.relative(testElmPackageDir, additionDir);
+    let absoluteSourceDirs = _.map(sourceDirs, p => this.util.resolveDir(testElmPackageDir, p));
+
+    let relativeSourceDirectories =
+      _.map(additions, p => path.join(relativePath, p)
+        .replace(/\\/g, "/"))
+        .filter(p => absoluteSourceDirs.indexOf(this.util.resolveDir(testElmPackageDir, p)) === -1);
+
+    return sourceDirs.concat(relativeSourceDirectories);
   }
 
   public isImprovedMinimumConstraint(dependency: string, candidate: string): boolean {
@@ -47,6 +69,54 @@ export class ElmPackageHelperImp implements ElmPackageHelper {
     return dependencyLowerBound[1] < candidateLowerBound[1]
       || dependencyLowerBound[2] < candidateLowerBound[2]
       || dependencyLowerBound[3] < candidateLowerBound[3];
+  }
+
+  public isNotExistingDependency(dependencies: string[][], candidate: string[]): boolean {
+    return !_.find(dependencies, x => {
+      return x[0] === candidate[0] && !this.isImprovedMinimumConstraint(x[1], candidate[1]);
+    });
+  }
+
+  public mergeDependencies(baseElmPackage: ElmPackageJson, testElmPackage: ElmPackageJson, testFramework: PluginTestFrameworkWithConfig)
+    : string[][] {
+    let dependencies: string[][] = _.toPairs(testElmPackage.dependencies);
+
+    if (baseElmPackage.dependencies) {
+      let baseDependencies: string[][] = _.toPairs(baseElmPackage.dependencies)
+        .filter((base: string[]) => this.isNotExistingDependency(dependencies, base));
+
+      dependencies = dependencies.concat(baseDependencies);
+    }
+
+    if (testFramework.config.dependencies) {
+      let testFrameworkDependencies = _.toPairs(testFramework.config.dependencies)
+        .filter((base) => this.isNotExistingDependency(dependencies, base));
+
+      dependencies = dependencies.concat(testFrameworkDependencies);
+    }
+
+    return dependencies;
+  }
+
+  public mergeSourceDirectories(baseElmPackage: ElmPackageJson, baseElmPackageDir: string, testElmPackage: ElmPackageJson,
+                                testElmPackageDir: string, testDir: string): string[] {
+    let sourceDirs: string[] = _.clone(testElmPackage.sourceDirectories);
+
+    if (!sourceDirs) {
+      sourceDirs = [];
+    }
+
+    if (sourceDirs.indexOf(".") === -1) {
+      sourceDirs.push(".");
+    }
+
+    if (sourceDirs.indexOf(testDir) === -1) {
+      sourceDirs.push(testDir);
+    }
+
+    sourceDirs = this.addSourceDirectories(baseElmPackage.sourceDirectories, baseElmPackageDir, testElmPackageDir, sourceDirs);
+
+    return sourceDirs;
   }
 
   public path(elmPackageJsonDirectory: string): string {
