+++ test/unit/lib/builder.test.ts
@@ -5,10 +5,8 @@ import rewire = require("rewire");
 import * as Sinon from "sinon";
 import * as SinonChai from "sinon-chai";
 import {Builder, BuilderImp, createBuilder} from "../../../lib/builder";
-import {Dependencies, ExecutionContext, LoboConfig, PluginTestFrameworkWithConfig} from "../../../lib/plugin";
+import {ExecutionContext, LoboConfig} from "../../../lib/plugin";
 import {Logger} from "../../../lib/logger";
-import {ElmPackageHelper, ElmPackageJson} from "../../../lib/elm-package-helper";
-import {Util} from "../../../lib/util";
 
 let expect = chai.expect;
 chai.use(SinonChai);
@@ -18,33 +16,22 @@ describe("lib builder", () => {
   let RewiredBuilder = rewire("../../../lib/builder");
   let builder: BuilderImp;
   let mockConfirm: Sinon.SinonStub;
-  let mockHelper: ElmPackageHelper;
   let mockLogger: Logger;
-  let mockUtil: Util;
   let mockReject: (error: Error) => void;
   let mockResolve: () => void;
-  let revertPrompt: () => void;
 
   beforeEach(() => {
     mockConfirm = Sinon.stub();
-    revertPrompt = RewiredBuilder.__set__({promptly: {confirm: mockConfirm}});
     let rewiredImp = RewiredBuilder.__get__("BuilderImp");
     mockLogger = <Logger> {};
     mockLogger.debug = Sinon.spy();
     mockLogger.error = Sinon.spy();
     mockLogger.info = Sinon.spy();
     mockLogger.trace = Sinon.spy();
-    mockHelper = <ElmPackageHelper> {isImprovedMinimumConstraint: Sinon.stub(), path: x => x, read: Sinon.stub(), write: Sinon.stub()};
-    mockUtil = <Util> {};
-    mockUtil.resolveDir = Sinon.spy();
-    builder = new rewiredImp(mockHelper, mockLogger, mockUtil);
+    builder = new rewiredImp(mockLogger);
 
     mockReject = Sinon.spy();
     mockResolve = Sinon.spy();
-  });
-
-  afterEach(() => {
-    revertPrompt();
   });
 
   describe("createBuilder", () => {
@@ -58,248 +45,30 @@ describe("lib builder", () => {
   });
 
   describe("build", () => {
-    it("should not call ensureElmPackageExists when config.noUpdate is true", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: true};
-      let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.stub();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.ensureElmPackageExists).not.to.have.been.called;
-      });
-    });
-
-    it("should not call syncTestElmPackage when config.noUpdate is true", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: true};
-      let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.stub();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.syncTestElmPackage).not.to.have.been.called;
-      });
-    });
-
-    it("should call installDependencies when config.noUpdate is true", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: true};
-      let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.stub();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.installDependencies).to.have.been.called;
-      });
-    });
-
-    it("should call make when config.noUpdate is true", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: true};
-      let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.stub();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.make).to.have.been.called;
-      });
-    });
-
-    it("should call ensureElmPackageExists with the supplied config", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: false};
-      let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.stub();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.ensureElmPackageExists).to.have.been.calledWith(config, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call ensureElmPackageExists with the base directory of '.' and location 'current'", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: false};
-      let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.stub();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.ensureElmPackageExists).to.have.been.calledWith(Sinon.match.any, ".", "current");
-      });
-    });
-
-    it("should call ensureElmPackageExists with the supplied test directory and location of 'test", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: false};
-      let context = <ExecutionContext> {config, testDirectory: "foo", testFile: "bar"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.spy();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.ensureElmPackageExists).to.have.been.calledWith(Sinon.match.any, "foo", "tests");
-      });
-    });
-
-    it("should call syncTestElmPackage with the supplied config", () => {
-      // arrange
-      let config = <LoboConfig> { noUpdate: false};
-      let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.spy();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.syncTestElmPackage).to.have.been.calledWith(config, Sinon.match.any, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call syncTestElmPackage with a base directory of '.'", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: false};
-      let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.spy();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.syncTestElmPackage).to.have.been.calledWith(Sinon.match.any, ".", Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call syncTestElmPackage with the supplied test directory", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: false};
-      let context = <ExecutionContext> {config, testDirectory: "foo", testFile: "bar"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.spy();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.syncTestElmPackage).to.have.been.calledWith(Sinon.match.any, Sinon.match.any, "foo", Sinon.match.any);
-      });
-    });
-
-    it("should call syncTestElmPackage with the supplied test file directory", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: false};
-      let context = <ExecutionContext> {config, testDirectory: "foo", testFile: "bar/baz"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.spy();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.syncTestElmPackage).to.have.been.calledWith(Sinon.match.any, Sinon.match.any, Sinon.match.any, "bar");
-      });
-    });
-
-    it("should call installDependencies with the supplied config", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: false};
-      let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.spy();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.installDependencies).to.have.been.calledWith(config, Sinon.match.any);
-      });
-    });
-
-    it("should call installDependencies with the supplied test directory", () => {
-      // arrange
-      let config = <LoboConfig> {noUpdate: false};
-      let context = <ExecutionContext> {config, testDirectory: "foo", testFile: "bar"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.spy();
-
-      // act
-      let actual = builder.build(context);
-
-      // assert
-      return actual.then(() => {
-        expect(builder.installDependencies).to.have.been.calledWith(Sinon.match.any, "foo");
-      });
-    });
-
     it("should call make with the supplied config", () => {
       // arrange
       let config = <LoboConfig> {noUpdate: false};
       let context = <ExecutionContext> {config, testFile: "foo"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.spy();
+      let mockMake = Sinon.stub();
+      builder.make = mockMake;
+      mockMake.resolves();
+
+      // act
+      let actual = builder.build(context);
+
+      // assert
+      return actual.then(() => {
+        expect(builder.make).to.have.been.calledWith(config, Sinon.match.any, Sinon.match.any, Sinon.match.any);
+      });
+    });
+
+    it("should call make with the supplied buildOutputFilePath", () => {
+      // arrange
+      let config = <LoboConfig> {noUpdate: false};
+      let context = <ExecutionContext> {config, testFile: "foo"};
+      let mockMake = Sinon.stub();
+      builder.make = mockMake;
+      mockMake.resolves();
 
       // act
       let actual = builder.build(context);
@@ -314,1508 +83,34 @@ describe("lib builder", () => {
       // arrange
       let config = <LoboConfig> {noUpdate: false};
       let context = <ExecutionContext> {config, buildOutputFilePath: "foo", testDirectory: "bar", testFile: "baz"};
-      builder.ensureElmPackageExists = Sinon.stub();
-      builder.syncTestElmPackage = Sinon.stub();
-      builder.installDependencies = Sinon.stub();
-      builder.make = Sinon.spy();
+      let mockMake = Sinon.stub();
+      builder.make = mockMake;
+      mockMake.resolves();
 
       // act
       let actual = builder.build(context);
 
       // assert
       return actual.then(() => {
-        expect(builder.make).to.have.been.calledWith(Sinon.match.any, Sinon.match.any, "bar", Sinon.match.any);
-      });
-    });
-  });
-
-  describe("ensureElmPackageExists", () => {
-    describe("shelljs test is true", () => {
-      let revertShellJs: () => void;
-
-      beforeEach(() => {
-        revertShellJs = RewiredBuilder.__set__({shelljs: {test: () => true}});
-      });
-
-      afterEach(() => {
-        revertShellJs();
-      });
-
-      it("should do nothing when elm-package.json already exists", () => {
-        // arrange
-        let config = <LoboConfig> {prompt: false};
-
-        // act
-        let actual = builder.ensureElmPackageExists(config, "foo", "bar");
-
-        // assert
-        return actual.finally(() => {
-          expect(actual.isResolved()).to.be.true;
-        });
+        expect(builder.make).to.have.been.calledWith(Sinon.match.any, "foo", Sinon.match.any, Sinon.match.any);
       });
     });
 
-    describe("shelljs test is false", () => {
-      let revertShellJs: () => void;
-
-      beforeEach(() => {
-        revertShellJs = RewiredBuilder.__set__({shelljs: {test: () => false}});
-      });
-
-      afterEach(() => {
-        revertShellJs();
-      });
-
-      it("should not prompt the user before running elm package install when config.prompt is false", () => {
-        // arrange
-        let config = <LoboConfig> {prompt: false};
-        builder.runElmPackageInstall = Sinon.spy((conf, testDir, prompt, resolve) => resolve());
-
-        // act
-        let actual = builder.ensureElmPackageExists(config, "foo", "bar");
-
-        // assert
-        return actual.finally(() => {
-          expect(mockConfirm).not.to.have.been.called;
-          expect(builder.runElmPackageInstall).to.have.been.calledWith();
-        });
-      });
-
-      it("should prompt the user before running elm package install when config.prompt is true", () => {
-        // arrange
-        let config = <LoboConfig> {prompt: true};
-        builder.runElmPackageInstall = Sinon.spy((conf, testDir, prompt, resolve) => resolve());
-        mockConfirm.callsFake((message, defaults, action) => action({}, "foo"));
-
-        // act
-        let actual = builder.ensureElmPackageExists(config, "foo", "bar");
-
-        // assert
-        return actual.catch(() => {
-          expect(mockConfirm).to.have.been.called;
-          expect(builder.runElmPackageInstall).not.to.have.been.called;
-        });
-      });
-
-      it("should not call runElmPackageInstall when config.prompt is true and error occurs", () => {
-        // arrange
-        let config = <LoboConfig> {prompt: true};
-        builder.runElmPackageInstall = Sinon.spy((conf, testDir, prompt, resolve) => resolve());
-        mockConfirm.callsFake((message, defaults, action) => action({}, "foo"));
-
-        // act
-        let actual = builder.ensureElmPackageExists(config, "foo", "bar");
-
-        // assert
-        return actual.catch(() => {
-          expect(builder.runElmPackageInstall).not.to.have.been.called;
-        });
-      });
-
-      it("should not call runElmPackageInstall when config.prompt is true and user answers false", () => {
-        // arrange
-        let config = <LoboConfig> {prompt: true};
-        builder.runElmPackageInstall = Sinon.spy((conf, testDir, prompt, resolve) => resolve());
-        mockConfirm.callsFake((message, defaults, action) => action(undefined, false));
-
-        // act
-        let actual = builder.ensureElmPackageExists(config, "foo", "bar");
-
-        // assert
-        return actual.catch(() => {
-          expect(builder.runElmPackageInstall).not.to.have.been.called;
-        });
-      });
-
-      it("should call runElmPackageInstall with prompt false when config.prompt is false", () => {
-        // arrange
-        let config = <LoboConfig> {prompt: false};
-        builder.runElmPackageInstall = Sinon.spy((conf, testDir, prompt, resolve) => resolve());
-
-        // act
-        let actual = builder.ensureElmPackageExists(config, "foo", "bar");
-
-        // assert
-        return actual.finally(() => {
-          expect(builder.runElmPackageInstall).to.have.been
-            .calledWith(Sinon.match.any, Sinon.match.any, false, Sinon.match.any, Sinon.match.any);
-        });
-      });
-
-      it("should call runElmPackageInstall with prompt true when config.prompt is true", () => {
-        // arrange
-        let config = <LoboConfig> {prompt: true};
-        builder.runElmPackageInstall = Sinon.spy((conf, testDir, prompt, resolve) => resolve());
-        mockConfirm.callsFake((message, defaults, action) => action(undefined, true));
-
-        // act
-        let actual = builder.ensureElmPackageExists(config, "foo", "bar");
-
-        // assert
-        return actual.finally(() => {
-          expect(builder.runElmPackageInstall).to.have.been
-            .calledWith(Sinon.match.any, Sinon.match.any, true, Sinon.match.any, Sinon.match.any);
-        });
-      });
-    });
-  });
-
-  describe("syncTestElmPackage", () => {
-    it("should call readElmPackage with the supplied base package directory", () => {
+    it("should call make with the supplied test file", () => {
       // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
+      let config = <LoboConfig> {noUpdate: false};
+      let context = <ExecutionContext> {config, buildOutputFilePath: "foo", testDirectory: "bar", testFile: "baz"};
+      let mockMake = Sinon.stub();
+      builder.make = mockMake;
+      mockMake.resolves();
 
       // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
+      let actual = builder.build(context);
 
       // assert
       return actual.then(() => {
-        expect(builder.readElmPackage).to.have.been.calledWith("bar", Sinon.match.any);
+        expect(builder.make).to.have.been.calledWith(Sinon.match.any, Sinon.match.any, Sinon.match.any, "baz");
       });
-    });
-
-    it("should call readElmPackage with the supplied test package directory", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.readElmPackage).to.have.been.calledWith(Sinon.match.any, "baz");
-      });
-    });
-
-    it("should call updateSourceDirectories with the supplied config", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateSourceDirectories).to.have.been
-          .calledWith(config, Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call updateSourceDirectories with the supplied base package directory", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateSourceDirectories)
-          .to.have.been.calledWith(Sinon.match.any, "bar", Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call updateSourceDirectories with the result.base from readElmPackage", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "abc", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateSourceDirectories)
-          .to.have.been.calledWith(Sinon.match.any, Sinon.match.any, "abc", Sinon.match.any, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call updateSourceDirectories with the supplied test package directory", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateSourceDirectories)
-          .to.have.been.calledWith(Sinon.match.any, Sinon.match.any, Sinon.match.any, "baz", Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call updateSourceDirectories with the supplied test dir", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateSourceDirectories)
-          .to.have.been.calledWith(Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any, "qux", Sinon.match.any);
-      });
-    });
-
-    it("should call updateSourceDirectories with the result.test from readElmPackage", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "abc"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateSourceDirectories)
-          .to.have.been.calledWith(Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any, "abc");
-      });
-    });
-
-    it("should call updateDependencies with the supplied config", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateDependencies).to.have.been.calledWith(config, Sinon.match.any, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call updateDependencies with the result.base from updateSourceDirectories", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "abc", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateDependencies).to.have.been.calledWith(Sinon.match.any, "abc", Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call updateDependencies with the supplied test directory", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "abc", test: "b"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateDependencies).to.have.been.calledWith(Sinon.match.any, Sinon.match.any, "baz", Sinon.match.any);
-      });
-    });
-
-    it("should call updateDependencies with the result.test from updateSourceDirectories", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      builder.readElmPackage = Sinon.stub();
-      (<Sinon.SinonStub>builder.readElmPackage).resolves({base: "a", test: "b"});
-      builder.updateSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateSourceDirectories).resolves({base: "a", test: "abc"});
-      builder.updateDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.updateDependencies).resolves({});
-
-      // act
-      let actual = builder.syncTestElmPackage(config, "bar", "baz", "qux");
-
-      // assert
-      return actual.then(() => {
-        expect(builder.updateDependencies).to.have.been.calledWith(Sinon.match.any, Sinon.match.any, Sinon.match.any, "abc");
-      });
-    });
-  });
-
-  describe("readElmPackage", () => {
-    it("should log an error when unable to read the source elm-package.json from the supplied base directory", () => {
-      // arrange
-      (<Sinon.SinonStub>mockHelper.read).onSecondCall().returns({});
-
-      // act
-      let actual = builder.readElmPackage("foo", "bar");
-
-      // assert
-      return actual.catch(() => {
-        expect(mockLogger.error).to.have.been.calledWith(Sinon.match(/main elm-package.json/));
-      });
-    });
-
-    it("should log an error when unable to read the test elm-package.json from the supplied test directory", () => {
-      // arrange
-      (<Sinon.SinonStub>mockHelper.read).onFirstCall().returns({});
-
-      // act
-      let actual = builder.readElmPackage("foo", "bar");
-
-      // assert
-      return actual.catch(() => {
-        expect(mockLogger.error).to.have.been.calledWith(Sinon.match(/test elm-package.json/));
-      });
-    });
-
-    it("should read the source elm-package.json from the supplied base directory", () => {
-      // arrange
-      (<Sinon.SinonStub>mockHelper.read).returns({});
-
-      // act
-      let actual = builder.readElmPackage("foo", "bar");
-
-      // assert
-      return actual.finally(() => {
-        expect(mockHelper.read).to.have.been.calledWith("foo");
-      });
-    });
-
-    it("should read the test elm-package.json from the supplied test directory", () => {
-      // arrange
-      (<Sinon.SinonStub>mockHelper.read).returns({});
-
-      // act
-      let actual = builder.readElmPackage("foo", "bar");
-
-      // assert
-      return actual.finally(() => {
-        expect(mockHelper.read).to.have.been.calledWith("bar");
-      });
-    });
-
-    it("should return the base json values", () => {
-      // arrange
-      let expectedSource = {name: "source"};
-      let expectedTest = {name: "test"};
-      (<Sinon.SinonStub>mockHelper.read).onFirstCall().returns(expectedSource);
-      (<Sinon.SinonStub>mockHelper.read).onSecondCall().returns(expectedTest);
-
-      // act
-      let actual = builder.readElmPackage("foo", "bar");
-
-      // assert
-      return actual.then((result: { base: {} }) => {
-        expect(result.base).to.equal(expectedSource);
-      });
-    });
-
-    it("should return the test json values", () => {
-      // arrange
-      let expectedSource = {name: "source"};
-      let expectedTest = {name: "test"};
-      (<Sinon.SinonStub>mockHelper.read).onFirstCall().returns(expectedSource);
-      (<Sinon.SinonStub>mockHelper.read).onSecondCall().returns(expectedTest);
-
-      // act
-      let actual = builder.readElmPackage("foo", "bar");
-
-      // assert
-      return actual.then((result: { test: {} }) => {
-        expect(result.test).to.equal(expectedTest);
-      });
-    });
-  });
-
-  describe("updateSourceDirectories", () => {
-    it("should call mergeSourceDirectories with the specified base package json", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.mergeSourceDirectories = Sinon.stub();
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.mergeSourceDirectories)
-          .to.have.been.calledWith(sourcePackageJson, Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call mergeSourceDirectories with the specified base directory", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.mergeSourceDirectories = Sinon.stub();
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.mergeSourceDirectories)
-          .to.have.been.calledWith(Sinon.match.any, "bar", Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call mergeSourceDirectories with the specified test package json", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.mergeSourceDirectories = Sinon.stub();
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.mergeSourceDirectories)
-          .to.have.been.calledWith(Sinon.match.any, Sinon.match.any, testPackageJson, Sinon.match.any, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call mergeSourceDirectories with the specified main test directory", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.mergeSourceDirectories = Sinon.stub();
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.mergeSourceDirectories)
-          .to.have.been.calledWith(Sinon.match.any, Sinon.match.any, Sinon.match.any, "baz", Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call mergeSourceDirectories with the specified test file directory", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.mergeSourceDirectories = Sinon.stub();
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.mergeSourceDirectories)
-          .to.have.been.calledWith(Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any, "qux", Sinon.match.any);
-      });
-    });
-
-    it("should call mergeSourceDirectories with the specified testFramework", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.mergeSourceDirectories = Sinon.stub();
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.mergeSourceDirectories).to.have.been
-          .calledWith(Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any, Sinon.match.any, config.testFramework);
-      });
-    });
-
-    it("should return the unaltered base package json when there is no difference", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.mergeSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.mergeSourceDirectories).returns(testPackageJson.sourceDirectories);
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.then((result: { base: {} }) => {
-        expect(result.base).to.equal(sourcePackageJson);
-      });
-    });
-
-    it("should return the unaltered test package json when there is no difference", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.mergeSourceDirectories = Sinon.stub();
-      (<Sinon.SinonStub>builder.mergeSourceDirectories).returns(testPackageJson.sourceDirectories);
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.then((result: { test: {} }) => {
-        expect(result.test).to.equal(testPackageJson);
-      });
-    });
-
-    it("should not prompt the user before running updating source directories when config.prompt is false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: false};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(mockConfirm).not.to.have.been.called;
-        expect(builder.updateSourceDirectoriesAction).to.have.been.calledWith();
-      });
-    });
-
-    it("should prompt the user before running updating source directories when config.prompt is true", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action({}, "foo"));
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.catch(() => {
-        expect(mockConfirm).to.have.been.called;
-        expect(builder.updateSourceDirectoriesAction).not.to.have.been.called;
-      });
-    });
-
-    it("should not call updateSourceDirectoriesAction when config.prompt is true and error occurs", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action({}, "foo"));
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.catch(() => {
-        expect(builder.updateSourceDirectoriesAction).not.to.have.been.called;
-      });
-    });
-
-    it("should not call updateSourceDirectoriesAction when config.prompt is true and user answers false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action(undefined, false));
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.catch(() => {
-        expect(builder.updateSourceDirectoriesAction).not.to.have.been.called;
-      });
-    });
-
-    it("should call updateSourceDirectoriesAction with merged sourceDirectories when config.prompt is false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: false};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-      builder.mergeSourceDirectories = Sinon.stub();
-      let expected = ["foo", "bar"];
-      (<Sinon.SinonStub>builder.mergeSourceDirectories).returns(expected);
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.updateSourceDirectoriesAction).to.have.been.calledWith(expected, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should return the unaltered base package json when there is a difference and config.prompt is false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: false};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.then((result: { base: {} }) => {
-        expect(result.base).to.equal(sourcePackageJson);
-      });
-    });
-
-    it("should return the updated test package json when there is a difference and config.prompt is false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: false};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-      let expected = <ElmPackageJson> {sourceDirectories: ["foo", "bar"]};
-      (<Sinon.SinonStub>builder.updateSourceDirectoriesAction).returns(expected);
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.then((result: { test: {} }) => {
-        expect(result.test).to.equal(expected);
-      });
-    });
-
-    it("should call updateSourceDirectoriesAction with merged sourceDirectories when config.prompt is true", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action(undefined, true));
-      builder.mergeSourceDirectories = Sinon.stub();
-      let expected = ["foo", "bar"];
-      (<Sinon.SinonStub>builder.mergeSourceDirectories).returns(expected);
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.updateSourceDirectoriesAction).to.have.been.calledWith(expected, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should return the unaltered base package json when there is a difference and config.prompt is true", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action(undefined, true));
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.then((result: { base: {} }) => {
-        expect(result.base).to.equal(sourcePackageJson);
-      });
-    });
-
-    it("should return the updated test package json when there is a difference and config.prompt is true", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-      builder.updateSourceDirectoriesAction = Sinon.stub();
-      let expected = <ElmPackageJson> {sourceDirectories: ["foo", "bar"]};
-      (<Sinon.SinonStub>builder.updateSourceDirectoriesAction).returns(expected);
-      mockConfirm.callsFake((message, defaults, action) => action(undefined, true));
-
-      // act
-      let actual = builder.updateSourceDirectories(config, "bar", sourcePackageJson, "baz", "qux", testPackageJson);
-
-      // assert
-      return actual.then((result: { test: {} }) => {
-        expect(result.test).to.equal(expected);
-      });
-    });
-  });
-
-  describe("updateSourceDirectoriesAction", () => {
-    it("should update the package json sourceDirectories with the supplied value", () => {
-      // arrange
-      let expected = ["foo"];
-
-      // act
-      let actual = builder.updateSourceDirectoriesAction(expected, "bar", <ElmPackageJson>{});
-
-      // assert
-      expect(actual.sourceDirectories).to.equal(expected);
-    });
-
-    it("should write the updated package json to the supplied directory", () => {
-      // act
-      builder.updateSourceDirectoriesAction(["foo"], "bar", <ElmPackageJson>{});
-
-      // assert
-      expect(mockHelper.write).to.have.been.calledWith("bar", Sinon.match.any);
-    });
-
-    it("should write the updated package json to the supplied directory", () => {
-      // arrange
-      let expected = ["foo"];
-
-      // act
-      builder.updateSourceDirectoriesAction(expected, "bar", <ElmPackageJson>{});
-
-      // assert
-      expect(mockHelper.write).to.have.been.calledWith(Sinon.match.any, {sourceDirectories: expected});
-    });
-  });
-
-  describe("updateDependencies", () => {
-    it("should call mergeDependencies with the specified base package json", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.mergeDependencies = Sinon.stub();
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.mergeDependencies)
-          .to.have.been.calledWith(sourcePackageJson, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should call mergeDependencies with the specified test package json", () => {
-      // arrange
-      let config = <LoboConfig> {prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.mergeDependencies = Sinon.stub();
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.mergeDependencies)
-          .to.have.been.calledWith(Sinon.match.any, testPackageJson, Sinon.match.any);
-      });
-    });
-
-    it("should call mergeDependencies with the specified testFramework", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.mergeDependencies = Sinon.stub();
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.mergeDependencies)
-          .to.have.been.calledWith(Sinon.match.any, Sinon.match.any, config.testFramework);
-      });
-    });
-
-    it("should return the unaltered base package json when there is no difference", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.mergeDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.mergeDependencies).returns(testPackageJson.dependencies);
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.then((result: { base: {} }) => {
-        expect(result.base).to.equal(sourcePackageJson);
-      });
-    });
-
-    it("should return the unaltered test package json when there is no difference", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.mergeDependencies = Sinon.stub();
-      (<Sinon.SinonStub>builder.mergeDependencies).returns(testPackageJson.dependencies);
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.then((result: { test: {} }) => {
-        expect(result.test).to.equal(testPackageJson);
-      });
-    });
-
-    it("should not prompt the user before running updating source directories when config.prompt is false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: false};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(mockConfirm).not.to.have.been.called;
-        expect(builder.updateDependenciesAction).to.have.been.calledWith();
-      });
-    });
-
-    it("should prompt the user before running updating source directories when config.prompt is true", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action({}, "foo"));
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.catch(() => {
-        expect(mockConfirm).to.have.been.called;
-        expect(builder.updateDependenciesAction).not.to.have.been.called;
-      });
-    });
-
-    it("should not call updateDependenciesAction when config.prompt is true and error occurs", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action({}, "foo"));
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.catch(() => {
-        expect(builder.updateDependenciesAction).not.to.have.been.called;
-      });
-    });
-
-    it("should not call updateDependenciesAction when config.prompt is true and user answers false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action(undefined, false));
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.catch(() => {
-        expect(builder.updateDependenciesAction).not.to.have.been.called;
-      });
-    });
-
-    it("should call updateDependenciesAction with merged dependencies when config.prompt is false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: false};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-      builder.mergeDependencies = Sinon.stub();
-      let expected = ["foo", "bar"];
-      (<Sinon.SinonStub>builder.mergeDependencies).returns(expected);
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.updateDependenciesAction).to.have.been.calledWith(expected, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should return the unaltered base package json when there is a difference and config.prompt is false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: false};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.then((result: { base: {} }) => {
-        expect(result.base).to.equal(sourcePackageJson);
-      });
-    });
-
-    it("should return the updated test package json when there is a difference and config.prompt is false", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: false};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-      let expected = <ElmPackageJson> {dependencies: <Dependencies>{foo: "qux"}};
-      (<Sinon.SinonStub>builder.updateDependenciesAction).returns(expected);
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.then((result: { test: {} }) => {
-        expect(result.test).to.equal(expected);
-      });
-    });
-
-    it("should call updateDependenciesAction with merged dependencies when config.prompt is true", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action(undefined, true));
-      builder.mergeDependencies = Sinon.stub();
-      let expected = ["foo", "bar"];
-      (<Sinon.SinonStub>builder.mergeDependencies).returns(expected);
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.updateDependenciesAction).to.have.been.calledWith(expected, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should return the unaltered base package json when there is a difference and config.prompt is true", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-      mockConfirm.callsFake((message, defaults, action) => action(undefined, true));
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.then((result: { base: {} }) => {
-        expect(result.base).to.equal(sourcePackageJson);
-      });
-    });
-
-    it("should return the updated test package json when there is a difference and config.prompt is true", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {dependencies: <Dependencies> {foo: "bar"}}}, prompt: true};
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "abc"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "def"}};
-      builder.updateDependenciesAction = Sinon.stub();
-      let expected = <ElmPackageJson> {dependencies: <Dependencies> {foo: "qux"}};
-      (<Sinon.SinonStub>builder.updateDependenciesAction).returns(expected);
-      mockConfirm.callsFake((message, defaults, action) => action(undefined, true));
-
-      // act
-      let actual = builder.updateDependencies(config, sourcePackageJson, "baz", testPackageJson);
-
-      // assert
-      return actual.then((result: { test: {} }) => {
-        expect(result.test).to.equal(expected);
-      });
-    });
-  });
-
-  describe("updateDependenciesAction", () => {
-    it("should update the package json dependencies with the supplied value", () => {
-      // arrange
-      let expected = [["foo", "bar"]];
-
-      // act
-      let actual = builder.updateDependenciesAction(expected, "baz", <ElmPackageJson>{});
-
-      // assert
-      expect(actual.dependencies).to.deep.equal({foo: "bar"});
-    });
-
-    it("should write the updated package json to the supplied directory", () => {
-      // act
-      builder.updateDependenciesAction([["foo", "bar"]], "baz", <ElmPackageJson>{});
-
-      // assert
-      expect(mockHelper.write).to.have.been.calledWith("baz", Sinon.match.any);
-    });
-
-    it("should write the updated package json to the supplied directory", () => {
-      // arrange
-      let expected = [["foo", "baz"]];
-
-      // act
-      builder.updateDependenciesAction(expected, "baz", <ElmPackageJson>{});
-
-      // assert
-      expect(mockHelper.write).to.have.been.calledWith(Sinon.match.any, Sinon.match(value => value.dependencies.foo = "bar"));
-    });
-  });
-
-  describe("mergeSourceDirectories", () => {
-    it("should return array with current dir only when no other dirs are specified", () => {
-      // arrange
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {sourceDirectories: []}};
-
-      // act
-      let actual = builder.mergeSourceDirectories(<ElmPackageJson>{}, "sourceDir", <ElmPackageJson>{}, "testDir", ".", testFramework);
-
-      // assert
-      expect(actual.length).to.equal(1);
-      expect(actual).to.include(".");
-    });
-
-    it("should return array with current dir and test file dir when no other dirs are specified", () => {
-      // arrange
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {sourceDirectories: []}};
-
-      // act
-      let actual = builder.mergeSourceDirectories(<ElmPackageJson>{}, "sourceDir", <ElmPackageJson>{}, "testDir", "foo", testFramework);
-
-      // assert
-      expect(actual.length).to.equal(2);
-      expect(actual).to.include(".");
-      expect(actual).to.include("foo");
-    });
-
-    it("should return array with current dir only when no other dirs are specified other than the test source directories", () => {
-      // arrange
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {sourceDirectories: []}};
-
-      // act
-      let actual = builder
-        .mergeSourceDirectories(<ElmPackageJson>{}, "sourceDir", <ElmPackageJson>{sourceDirectories: ["."]}, "testDir", ".", testFramework);
-
-      // assert
-      expect(actual.length).to.equal(1);
-      expect(actual).to.include(".");
-    });
-
-    it("should return array with current dir relative test directory", () => {
-      // arrange
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {sourceDirectories: ["foo"]}};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-
-      // act
-      let actual = builder.mergeSourceDirectories(sourcePackageJson, "sourceDir", testPackageJson, "testDir", "qux", testFramework);
-
-      // assert
-      expect(actual).to.include(".");
-    });
-
-    it("should return array with test directory relative test directory", () => {
-      // arrange
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {sourceDirectories: ["foo"]}};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-
-      // act
-      let actual = builder.mergeSourceDirectories(sourcePackageJson, "sourceDir", testPackageJson, "testDir", "qux", testFramework);
-
-      // assert
-      expect(actual).to.include("test");
-    });
-
-    it("should return array with base source directories relative test directory", () => {
-      // arrange
-      mockUtil.resolveDir = (...dirs) => dirs.join();
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {sourceDirectories: ["foo"]}};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-
-      // act
-      let actual = builder.mergeSourceDirectories(sourcePackageJson, "sourceDir", testPackageJson, "testDir", "qux", testFramework);
-
-      // assert
-      expect(actual).to.include("../sourceDir/source");
-    });
-
-    it("should return array with lobo relative test directory", () => {
-      // arrange
-      mockUtil.resolveDir = (...dirs) => dirs.join();
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {sourceDirectories: ["foo"]}};
-      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
-      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
-
-      // act
-      let actual = builder.mergeSourceDirectories(sourcePackageJson, "sourceDir", testPackageJson, "testDir", "qux", testFramework);
-
-      // assert
-      expect(actual).to.include.something.that.match(/..\/foo/);
-    });
-  });
-
-  describe("addSourceDirectories", () => {
-    it("should return the unaltered source directories when the additions does not exist", () => {
-      // arrange
-      let expected = ["abc"];
-
-      // act
-      let actual = builder.addSourceDirectories(undefined, "foo", "bar", expected);
-
-      // assert
-      expect(actual).to.equal(expected);
-    });
-
-    it("should return directories with added additions relative to the test directory when directories are same", () => {
-      // arrange
-      mockUtil.resolveDir = (...dirs) => dirs.join();
-
-      // act
-      let actual = builder.addSourceDirectories(["foo"], "bar", "bar", ["qux"]);
-
-      // assert
-      expect(actual).to.include("qux");
-      expect(actual).to.include("foo");
-    });
-
-    it("should return directories with added additions relative to the test directory when directories are different", () => {
-      // arrange
-      mockUtil.resolveDir = (...dirs) => dirs.join();
-
-      // act
-      let actual = builder.addSourceDirectories(["foo"], "bar", "baz", ["qux"]);
-
-      // assert
-      expect(actual).to.include("qux");
-      expect(actual).to.include("../bar/foo");
-    });
-
-    it("should return directories with added additions relative to the test directory when test directory is sub-directory", () => {
-      // arrange
-      mockUtil.resolveDir = (...dirs) => dirs.join();
-
-      // act
-      let actual = builder.addSourceDirectories(["foo"], "bar", "bar/baz", ["qux"]);
-
-      // assert
-      expect(actual).to.include("qux");
-      expect(actual).to.include("../foo");
-    });
-  });
-
-  describe("mergeDependencies", () => {
-    it("should return empty dependency list when source, test and framework dependencies do not exist", () => {
-      // arrange
-      let sourcePackageJson = <ElmPackageJson>{};
-      let testPackageJson = <ElmPackageJson>{};
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {}};
-
-      // act
-      let actual = builder.mergeDependencies(sourcePackageJson, testPackageJson, testFramework);
-
-      // assert
-      expect(actual.length).to.equal(0);
-    });
-
-    it("should return the source and test framework dependencies when the test dependencies does not exist", () => {
-      // arrange
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "foo"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> undefined};
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {dependencies: <Dependencies> {framework: "baz"}}};
-
-      // act
-      let actual = builder.mergeDependencies(sourcePackageJson, testPackageJson, testFramework);
-
-      // assert
-      expect(actual.length).to.equal(2);
-      expect(actual).to.include.something.deep.equal(["source", "foo"]);
-      expect(actual).to.include.something.deep.equal(["framework", "baz"]);
-    });
-
-    it("should return the source test and test framework dependencies for the supplied parameters", () => {
-      // arrange
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "foo"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "bar"}};
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {dependencies: <Dependencies> {framework: "baz"}}};
-
-      // act
-      let actual = builder.mergeDependencies(sourcePackageJson, testPackageJson, testFramework);
-
-      // assert
-      expect(actual.length).to.equal(3);
-      expect(actual).to.include.something.deep.equal(["source", "foo"]);
-      expect(actual).to.include.something.deep.equal(["framework", "baz"]);
-    });
-
-    it("should return the source test and test framework without duplicates", () => {
-      // arrange
-      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {foo: "bar"}};
-      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {foo: "bar"}};
-      let testFramework = <PluginTestFrameworkWithConfig> {config: {dependencies: <Dependencies> {foo: "bar"}}};
-
-      // act
-      let actual = builder.mergeDependencies(sourcePackageJson, testPackageJson, testFramework);
-
-      // assert
-      expect(actual.length).to.equal(1);
-      expect(actual).to.include.something.deep.equal(["foo", "bar"]);
-    });
-  });
-
-  describe("isNotExistingDependency", () => {
-    it("should return false when the candidate exists in the dependencies", () => {
-      // act
-      let actual = builder.isNotExistingDependency([["foo", "bar"]], ["foo", "bar"]);
-
-      // assert
-      expect(actual).to.be.false;
-    });
-
-    it("should return true when the candidate does not exist in the dependencies", () => {
-      // act
-      let actual = builder.isNotExistingDependency([["foo", "bar"]], ["baz", "qux"]);
-
-      // assert
-      expect(actual).to.be.true;
-    });
-
-    it("should return true when the candidate is an improved constraint", () => {
-      // arrange
-      mockHelper.isImprovedMinimumConstraint = () => true;
-
-      // act
-      let actual = builder.isNotExistingDependency([["foo", "1.0.0"]], ["foo", "2.0.0"]);
-
-      // assert
-      expect(actual).to.be.true;
-    });
-
-    it("should return false when the candidate is not an improved constraint", () => {
-      // arrange
-      mockHelper.isImprovedMinimumConstraint = () => false;
-
-      // act
-      let actual = builder.isNotExistingDependency([["foo", "2.0.0"]], ["foo", "1.0.0"]);
-
-      // assert
-      expect(actual).to.be.false;
-    });
-  });
-
-  describe("installDependencies", () => {
-    it("should return a promise that calls runElmPackageInstall with the supplied config", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}};
-      builder.runElmPackageInstall = Sinon.spy((conf, testDir, prompt, resolve) => resolve());
-
-      // act
-      let actual = builder.installDependencies(config, "bar");
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.runElmPackageInstall).to.have.been.calledWith(config, Sinon.match.any, Sinon.match.any);
-      });
-    });
-
-    it("should return a promise that calls runElmPackageInstall with the supplied config", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}};
-      builder.runElmPackageInstall = Sinon.spy((conf, testDir, prompt, resolve) => resolve());
-
-      // act
-      let actual = builder.installDependencies(config, "bar");
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.runElmPackageInstall).to.have.been.calledWith(Sinon.match.any, "bar", Sinon.match.any);
-      });
-    });
-
-    it("should return a promise that calls runElmPackageInstall with the supplied config.prompt", () => {
-      // arrange
-      let config = <LoboConfig> {testFramework: {config: {sourceDirectories: ["foo"]}}, prompt: true};
-      builder.runElmPackageInstall = Sinon.spy((conf, testDir, prompt, resolve) => resolve());
-
-      // act
-      let actual = builder.installDependencies(config, "bar");
-
-      // assert
-      return actual.finally(() => {
-        expect(builder.runElmPackageInstall).to.have.been.calledWith(Sinon.match.any, Sinon.match.any, true);
-      });
-    });
-  });
-
-  describe("runElmPackageInstall", () => {
-    let revertChildProcess: () => void;
-    let mockExec: Sinon.SinonStub;
-
-    beforeEach(() => {
-      mockExec = Sinon.stub();
-      mockResolve = Sinon.spy();
-      mockReject = Sinon.spy();
-      revertChildProcess = RewiredBuilder.__set__({childProcess: {execSync: mockExec}});
-    });
-
-    afterEach(() => {
-      revertChildProcess();
-    });
-
-    it("should not call elm-package to install the packages when config.noInstall is true", () => {
-      // arrange
-      let config = <LoboConfig> {compiler: "foo", noInstall: true};
-
-      // act
-      builder.runElmPackageInstall(config, "bar", true, mockResolve, mockReject);
-
-      // assert
-      expect(mockExec).not.to.have.been.called;
-    });
-
-    it("should call elm-package to install the packages", () => {
-      // arrange
-      let config = <LoboConfig> {compiler: "foo"};
-
-      // act
-      builder.runElmPackageInstall(config, "bar", true, mockResolve, mockReject);
-
-      // assert
-      expect(mockExec).to.have.been.calledWith(Sinon.match(/elm-package install/), Sinon.match.any);
-    });
-
-    it("should call elm-package to install the packages from the specified elm-install path", () => {
-      // arrange
-      let config = <LoboConfig> {compiler: "foo"};
-
-      // act
-      builder.runElmPackageInstall(config, "bar", true, mockResolve, mockReject);
-
-      // assert
-      expect(mockExec).to.have.been.calledWith(Sinon.match(/^foo([\/\\])elm-package install/), Sinon.match.any);
-    });
-
-    it("should call elm-package to install the packages without --yes when prompt is true", () => {
-      // arrange
-      let config = <LoboConfig> {compiler: "foo"};
-
-      // act
-      builder.runElmPackageInstall(config, "bar", true, mockResolve, mockReject);
-
-      // assert
-      expect(mockExec).to.have.been.calledWith(Sinon.match((x) => x.indexOf("--yes") === -1), Sinon.match.any);
-    });
-
-    it("should call elm-package to install the packages with --yes when prompt is false", () => {
-      // arrange
-      let config = <LoboConfig> {compiler: "foo"};
-
-      // act
-      builder.runElmPackageInstall(config, "bar", false, mockResolve, mockReject);
-
-      // assert
-      expect(mockExec).to.have.been.calledWith(Sinon.match(/ --yes/), Sinon.match.any);
-    });
-
-    it("should call elm-package to install the packages with cwd as the supplied directory", () => {
-      // arrange
-      let config = <LoboConfig> {compiler: "foo"};
-
-      // act
-      builder.runElmPackageInstall(config, "bar", false, mockResolve, mockReject);
-
-      // assert
-      expect(mockExec).to.have.been.calledWith(Sinon.match.any, Sinon.match((x => x.cwd === "bar")));
-    });
-
-    it("should call resolve when there are no elm-package install errors", () => {
-      // arrange
-      let config = <LoboConfig> {compiler: "foo"};
-
-      // act
-      builder.runElmPackageInstall(config, "bar", false, mockResolve, mockReject);
-
-      // assert
-      expect(mockResolve).to.have.been.calledWith();
-    });
-
-    it("should catch any elm-package installation errors and call the specified reject with the error", () => {
-      // arrange
-      let config = <LoboConfig> {};
-      let expected = new Error();
-      mockExec.throws(expected);
-
-      // act
-      builder.runElmPackageInstall(config, "bar", true, mockResolve, mockReject);
-
-      // assert
-      expect(mockReject).to.have.been.calledWith(expected);
     });
   });
