+++ test/unit/lib/elm-package-helper.test.ts
@@ -10,6 +10,7 @@ import {createElmPackageHelper, ElmPackageHelper, ElmPackageHelperImp, ElmPackag
 import {Logger} from "../../../lib/logger";
 import {Util} from "../../../lib/util";
 import {SinonStub} from "sinon";
+import {Dependencies, PluginTestFrameworkWithConfig} from "../../../lib/plugin";
 
 let expect = chai.expect;
 chai.use(SinonChai);
@@ -28,6 +29,7 @@ describe("lib elm-package-helper", () => {
     mockLogger.debug = Sinon.stub();
     mockRead = Sinon.stub();
     mockUtil = <Util><{}> { read: mockRead };
+    mockUtil.resolveDir = Sinon.spy();
     helper = new rewiredImp(mockLogger, mockUtil);
   });
 
@@ -38,6 +40,55 @@ describe("lib elm-package-helper", () => {
 
       // assert
       expect(actual).to.exist;
+    });
+  });
+
+  describe("addSourceDirectories", () => {
+    it("should return the unaltered source directories when the additions does not exist", () => {
+      // arrange
+      let expected = ["abc"];
+
+      // act
+      let actual = helper.addSourceDirectories(undefined, "foo", "bar", expected);
+
+      // assert
+      expect(actual).to.equal(expected);
+    });
+
+    it("should return directories with added additions relative to the test directory when directories are same", () => {
+      // arrange
+      mockUtil.resolveDir = (...dirs) => dirs.join();
+
+      // act
+      let actual = helper.addSourceDirectories(["foo"], "bar", "bar", ["qux"]);
+
+      // assert
+      expect(actual).to.include("qux");
+      expect(actual).to.include("foo");
+    });
+
+    it("should return directories with added additions relative to the test directory when directories are different", () => {
+      // arrange
+      mockUtil.resolveDir = (...dirs) => dirs.join();
+
+      // act
+      let actual = helper.addSourceDirectories(["foo"], "bar", "baz", ["qux"]);
+
+      // assert
+      expect(actual).to.include("qux");
+      expect(actual).to.include("../bar/foo");
+    });
+
+    it("should return directories with added additions relative to the test directory when test directory is sub-directory", () => {
+      // arrange
+      mockUtil.resolveDir = (...dirs) => dirs.join();
+
+      // act
+      let actual = helper.addSourceDirectories(["foo"], "bar", "bar/baz", ["qux"]);
+
+      // assert
+      expect(actual).to.include("qux");
+      expect(actual).to.include("../foo");
     });
   });
 
@@ -115,6 +166,172 @@ describe("lib elm-package-helper", () => {
     });
   });
 
+  describe("isNotExistingDependency", () => {
+    it("should return false when the candidate exists in the dependencies", () => {
+      // act
+      let actual = helper.isNotExistingDependency([["foo", "bar"]], ["foo", "bar"]);
+
+      // assert
+      expect(actual).to.be.false;
+    });
+
+    it("should return true when the candidate does not exist in the dependencies", () => {
+      // act
+      let actual = helper.isNotExistingDependency([["foo", "bar"]], ["baz", "qux"]);
+
+      // assert
+      expect(actual).to.be.true;
+    });
+
+    it("should return true when the candidate is an improved constraint", () => {
+      // arrange
+      helper.isImprovedMinimumConstraint = () => true;
+
+      // act
+      let actual = helper.isNotExistingDependency([["foo", "1.0.0"]], ["foo", "2.0.0"]);
+
+      // assert
+      expect(actual).to.be.true;
+    });
+
+    it("should return false when the candidate is not an improved constraint", () => {
+      // arrange
+      helper.isImprovedMinimumConstraint = () => false;
+
+      // act
+      let actual = helper.isNotExistingDependency([["foo", "2.0.0"]], ["foo", "1.0.0"]);
+
+      // assert
+      expect(actual).to.be.false;
+    });
+  });
+
+  describe("mergeDependencies", () => {
+    it("should return empty dependency list when source, test and framework dependencies do not exist", () => {
+      // arrange
+      let sourcePackageJson = <ElmPackageJson>{};
+      let testPackageJson = <ElmPackageJson>{};
+      let testFramework = <PluginTestFrameworkWithConfig> {config: {}};
+
+      // act
+      let actual = helper.mergeDependencies(sourcePackageJson, testPackageJson, testFramework);
+
+      // assert
+      expect(actual.length).to.equal(0);
+    });
+
+    it("should return the source and test framework dependencies when the test dependencies does not exist", () => {
+      // arrange
+      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "foo"}};
+      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> undefined};
+      let testFramework = <PluginTestFrameworkWithConfig> {config: {dependencies: <Dependencies> {framework: "baz"}}};
+
+      // act
+      let actual = helper.mergeDependencies(sourcePackageJson, testPackageJson, testFramework);
+
+      // assert
+      expect(actual.length).to.equal(2);
+      expect(actual).to.include.something.deep.equal(["source", "foo"]);
+      expect(actual).to.include.something.deep.equal(["framework", "baz"]);
+    });
+
+    it("should return the source test and test framework dependencies for the supplied parameters", () => {
+      // arrange
+      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {source: "foo"}};
+      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {test: "bar"}};
+      let testFramework = <PluginTestFrameworkWithConfig> {config: {dependencies: <Dependencies> {framework: "baz"}}};
+
+      // act
+      let actual = helper.mergeDependencies(sourcePackageJson, testPackageJson, testFramework);
+
+      // assert
+      expect(actual.length).to.equal(3);
+      expect(actual).to.include.something.deep.equal(["source", "foo"]);
+      expect(actual).to.include.something.deep.equal(["framework", "baz"]);
+    });
+
+    it("should return the source test and test framework without duplicates", () => {
+      // arrange
+      let sourcePackageJson = <ElmPackageJson>{dependencies: <Dependencies> {foo: "bar"}};
+      let testPackageJson = <ElmPackageJson>{dependencies: <Dependencies> {foo: "bar"}};
+      let testFramework = <PluginTestFrameworkWithConfig> {config: {dependencies: <Dependencies> {foo: "bar"}}};
+
+      // act
+      let actual = helper.mergeDependencies(sourcePackageJson, testPackageJson, testFramework);
+
+      // assert
+      expect(actual.length).to.equal(1);
+      expect(actual).to.include.something.deep.equal(["foo", "bar"]);
+    });
+  });
+
+  describe("mergeSourceDirectories", () => {
+    it("should return array with current dir only when no other dirs are specified", () => {
+      // act
+      let actual = helper.mergeSourceDirectories(<ElmPackageJson>{}, "sourceDir", <ElmPackageJson>{}, "testDir", ".");
+
+      // assert
+      expect(actual.length).to.equal(1);
+      expect(actual).to.include(".");
+    });
+
+    it("should return array with current dir and test file dir when no other dirs are specified", () => {
+      // act
+      let actual = helper.mergeSourceDirectories(<ElmPackageJson>{}, "sourceDir", <ElmPackageJson>{}, "testDir", "foo");
+
+      // assert
+      expect(actual.length).to.equal(2);
+      expect(actual).to.include(".");
+      expect(actual).to.include("foo");
+    });
+
+    it("should return array with current dir only when no other dirs are specified other than the test source directories", () => {
+      // act
+      let actual = helper
+        .mergeSourceDirectories(<ElmPackageJson>{}, "sourceDir", <ElmPackageJson>{sourceDirectories: ["."]}, "testDir", ".");
+
+      // assert
+      expect(actual.length).to.equal(1);
+      expect(actual).to.include(".");
+    });
+
+    it("should return array with current dir relative test directory", () => {
+      // arrange
+      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
+      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
+
+      // act
+      let actual = helper.mergeSourceDirectories(sourcePackageJson, "sourceDir", testPackageJson, "testDir", "qux");
+
+      // assert
+      expect(actual).to.include(".");
+    });
+
+    it("should return array with test directory relative test directory", () => {
+      // arrange
+      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
+      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
+
+      // act
+      let actual = helper.mergeSourceDirectories(sourcePackageJson, "sourceDir", testPackageJson, "testDir", "qux");
+
+      // assert
+      expect(actual).to.include("test");
+    });
+
+    it("should return array with base source directories relative test directory", () => {
+      // arrange
+      mockUtil.resolveDir = (...dirs) => dirs.join();
+      let sourcePackageJson = <ElmPackageJson>{sourceDirectories: ["source"]};
+      let testPackageJson = <ElmPackageJson>{sourceDirectories: ["test"]};
+
+      // act
+      let actual = helper.mergeSourceDirectories(sourcePackageJson, "sourceDir", testPackageJson, "testDir", "qux");
+
+      // assert
+      expect(actual).to.include("../sourceDir/source");
+    });
+  });
 
   describe("path", () => {
     it("should return path starting in supplied directory", () => {
